+++ joulescope/entry_points/runner.py
@@ -20,11 +20,11 @@ import sys
 import argparse
 import logging
 from joulescope.entry_points import bootloader_go, capture, capture_usb, \
-    gpo_demo, info, parameter_set, program, recording, stream_test
+    gpo_demo, info, parameter_set, program, recording, statistics, stream_test
 
 
 entry_points = [bootloader_go, capture, capture_usb, gpo_demo, info,
-                parameter_set, program, recording, stream_test]
+                parameter_set, program, recording, statistics, stream_test]
 """This list of available command modules.  Each module must contain a 
 parser_config(subparser) function.  The function must return the callable(args)
 that will be executed for the command."""